linux-yocto-3.10: ARM: Use Kernel fragments
This patch adds configuration fragments for linux ARM 3.10 standard
and preempt-rt kernels.
When compiling the kernel from the linux-yocto repository, the
fragments are used to build the defconfig depending on options such as
endianess, standard/preempt-rt, power management, debug and test
capabilities.
The fragments present in this meta-lsi yocto layer include only
hardware and software options that are not platform-dependent. The
fragments with Axxia ARM platform-specific options are taken from the
meta branch of the linux-yocto-3.10 repository.

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>