linux-yocto-3.4: ARM standard kernel fragments
This patch adds configuration fragments for linux ARM 3.4 standard
kernel.
The fragments are used to build the defconfig when compiling the
standard kernel from the linux-yocto-3.4 repository.
The fragments present in this meta-lsi yocto layer include only
hardware and software options that are not platform-dependent. The
fragments containing Axxia ARM platform-specific options are taken
from the meta branch of the linux-yocto-3.4 repository.

Signed-off-by: Cristian Bercaru <cristian.bercaru@windriver.com>