README: add information about configuration fragments
When building the kernel from the Yocto repositories, configuration
fragments are used to build the final .config instead of pre-built
defconfigs.
This patch adds information to the README about which variables need
to be set in local.conf in order to include certain configuration
fragments in the final .config. The fragments contain power-management
endianess and debugging options.

Signed-off-by: Cristian Bercaru <cristian.bercaru@windriver.com>