linux-yocto-3.14: add configuration fragments
This patch adds configuration fragments for linux 3.14 standard and
preempt-rt kernels.
When compiling the kernel from the linux-yocto repository, the
fragments are used to build the defconfig depending on options such as
capabilities.
The fragments present in this meta-lsi yocto layer include only
hardware and software options that are not platform-dependent. The
fragments with Axxia platform-specific options are taken from the meta
branch of the linux-yocto-3.14 repository.

Signed-off-by: Cristian Bercaru <cristian.bercaru@windriver.com>