linux-yocto-3.4: SMP config option becomes optional
By default, the SMP Kconfig option is enabled in linux-yocto-3.4 on
Axxia platforms. However, you can disable it by adding the following
line to 'local.conf':    SMP = "no"
This patch puts the SMP option in separate kernel configuration
fragment and adds the SMP option to enable or to disable it.
SMP defaults to "yes".

Signed-off-by: Cristian Bercaru <cristian.bercaru@windriver.com>