linux-yocto_3.14: Reset LSI_SRC if lsi-public is set
3.14 Github Kernel didn't have public version. Even if kernel version
is set to 3.10, bitbake will check also the 3.14 recipe and will test
if "require ${LSI_SRC}-rt_3.14.inc" can be be satisfied.
If lsi-public is set for 3.10 kernel, we'll get this error:
ERROR: Could not include required file lsi-public_3.14.inc

Reset LSI_SRC to "linux-yocto" in 3.14 recipe in case if LSI_SRC
is set to "lsi-public".

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>