rpm: Backport 5.4.16 version from master to work with db_6.2.23
ARMv8 tunes are supported just for a newer version of DB that cannot
work with the actual version of rpm since there are some changes in
db locking semantics.
Untill a new version of rpm will be released with the fix, use as a
workaround the newer version of rpm from master branch adapted for the
new db.

NOTE: This rpm version didn't contain a fix, just a workaround.
http://lists.openembedded.org/pipermail/openembedded-core/2016-June/122456.html

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>