netcf: Backport recipe from krogoth to include some fixes
There is a fix upstream for netcf which was not included in jethro
to fix the configure error:
error: pathspec 'gnulib' did not match any file(s) known to git.

Solution for this is to call bootstrap script with --no-git flag,
but this call is done in autotools-bootstrap bblass which was
inherited. Ro remove this inheritance from a bbappend is inposible,
so, the only solution to fix this from a bbappend is to copy the entire
recipe in meta-axxia.

Patch from krogoth branch with the fix:

commit 9199968550c1eb24cbf3b16798895091985953bb
netcf: fix mishandling of gnulib submodule causing build fail
Author:     Paul Gortmaker <paul.gortmaker@windriver.com>

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>