rmc: Port rmc and rmc-db recipes from meta-intel
RMC (Runtime Machine Configuration) provides a tool and libraries
to identify types of hardware boards and access any file-based data
specific to the board's type at runtime in a centralized way.

RMC-DB provides a centralized RMC database for RMC feature.

This is done in a porting effort of meta-intel metadata to meta-axxia.
The state of meta-intel is:
commit 6add41510412ca196efb3e4f949d403a8b6f35d7
linux-yocto/4.8: Update to latest stable release

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>