ypbind-mt: Set domainname and add -broadcast option
If NIS/YP is not fully configured, it will fail on starting by systemd.
Set a generic domain name which will need to be changed with a working
one.

[FAILED] Failed to start YP Bind.
ypbind[1946]: domainname not set - aborting.

Also, add -broadcast flag on /usr/sbin/ypbind call to only "send a
broadcast to request the information needed to bind to a specific NIS
server. With this option, /etc/yp.conf will be ignored." (man ypbind)

ypbind[2208]: No NIS server and no -broadcast option specified.
ypbind[2208]: Add a NIS server to the /etc/yp.conf configuration file,
ypbind[2208]: or start ypbind with the -broadcast option.

After NIS/YP is configured, the service need to be restarted:
$ systemctl restart ypbind.service

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>