linux-yocto: BBMASK kernel bbappend recipes from meta-intel
Because meta-axxia inherit a lot of metadata from meta-intel,
especially axxiax86-64 machine include all the matadata
from intel-corei7-64 machine, we inherit also the preferred
versions of kernel from meta-intel.

This patch add all the kernel bbappend recipes to BBMASK list
to be skip by bitbake (as we did for meta-virtualization).

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>