linux: Add u-boot and dtc as dependencies
Starting with pyro branch (2.3), a new mechanism is used
for sysroot in OE: Recipe Specific Sysroots

from https://patchwork.openembedded.org/patch/136198:
"Switching the system to build using recipe specific sysroots and
where changes could be isolated from it, that has been done.

With the current single sysroot approach, its possible for software
to find things which aren't in their dependencies. This leads to a
determinism problem and is a growing issue in several of the market
segments where OE makes sense. The way to solve this problem for OE
is to have seperate sysroots for each recipe and these will only
contain the dependencies for that recipe."

So, specific sysroot directory is used for linux recipe and others
binaries which are not in this dir are ignored.
For compiling fit images we are using mkimage and dtc which were
installed on host untill morty branch. Now, we need to compile and
add them as depencencies for linux to be installed in the specific
sysroot for linux package. Otherwise, those binaries from host will
no longer be accessible by bitbake:

| mkimage: command not found
| sh: 1: dtc: not found
| /usr/bin/mkimage: Can't read linux.fit.tmp: Invalid argument
| /usr/bin/mkimage Can't add hashes to FIT blob
| GNUmakefile:62: recipe for target 'linux.fit' failed
| make: *** [linux.fit] Error 255

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>