nfs-utils: Fix nfs-mountd and nfs-server services
nfs-mountd and nfs-server services were inactive (dead) on boot
because a condition from the service files was not met:
ConditionPathExists=/etc/exports was not met

By creating and installing an empty exports file in rootfs, the
services are active. /etc/exports will be populated by the final user
with the access control list for filesystems which may be exported.

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>