kernel-devsrc: Add missing files for Makefile "clean" target
Starting with commit 9af0f1a46bbb6ad9ee8b35957251f4aa826b023f
("kernel-devsrc: restructure for out of tree (and on target) module
builds") from poky repo, the way how kernel-devsrc package is
created was changed.

The old kernel-devsrc package starts with a full copy of the kernel
source and then starts to strip out elements that are not required.

This restructuring of the package uses an approach similar to other
distros, where the kernel-devsrc package is for building against the
running kernel and uses a curated set of copied infrastructure, versus
a mass copy of the entire kernel.

More details about this change here:
https://patches.openembedded.org/patch/152442

This patch add in kernel-devsrc some Makefiles, Kbuild and .mk files
required to perform the "clean" target.

A0261 - After Installing the SDK, 'make clean oldconfig scripts
        tools/objtool' Fails

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>