linux-axxia: use defconfigs from kernel tree
In order to get rid of the full-defconfigs from the meta-layer which
should be syncronized permanently with the defconfigs from the kernel
source tree, let bitbake get automatically the defconfig from the kernel
tree on do_kernel_metadata build step.

Another big advantage of this change is that we can now build the debug
configs for meta-axxia as well. DBG variable will influence the chose
of the proper defconfig from kernel source tree.

A0219 - meta-axxia: linux-axxia kernel to use defconfigs from kernel tree

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>