dhclient: Fix resolv.conf if interface is used for nfsroot
If the main interface is used for nfsroot, dhclient skip it
by calling dhclient-systemd-wrapper. In this case, resolv.conf
continue to be maintained by systemd-resolvd service which is
not populating the file with nameservers from dhcp lease.

In the kernel official documentation [1], it is recomanded to
symlink /etc/resolv.conf to /proc/net/pnp on systems that use
an NFS root filesystem.

So, we append dhclient-systemd-wrapper to create that symlink
if nfsroot flag is different than default value (0).

[1] https://www.kernel.org/doc/Documentation/filesystems/nfs/nfsroot.txt

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>