linux: Fix LINUX_KERNEL_TYPE value inheritance
When FILESEXTRAPATHS_prepend is parsed and his value contains
LINUX_KERNEL_TYPE variable which is set after FILESEXTRAPATHS,
it will use the value defined upstream. So, for preempt-rt
recipes, it will use the upstream default which is "standard"
and it will cause problems on finding files from SRC_URI which
are in paths containing LINUX_KERNEL_TYPE. This warning and
errors occurs:

Unable to get checksum for linux-yocto-rt SRC_URI
	entry preempt-rt.scc: file could not be found
Unable to get checksum for linux-yocto-rt-debug SRC_URI
	entry preempt-rt.scc: file could not be found

Fix this by setting the value for LINUX_KERNEL_TYPE prior to
inherit linux-yocto.inc which add upstream default settings.

Signed-off-by: Daniel Dragomir <Daniel.Dragomir@windriver.com>