linux: Run depmod for alternative kernels
By default, do_rootfs() runs depmod on the final rootfs only for the
preferred virtual/kernel, so the depmod files for the main kernel are
in the rootfs image.

Run depmod for each alternative kernek to add the output of depmod into
rootfs for all the alternative kernels.

Also, the output of depmod will be included in each kernel modules .tgz.

A side effect is that lttng-modules are build only against the preferred
virtual/kernel, so those modules will not be availabe for alternative kernels.

B0179 - Only the "Main" Kernel has Modules in the RFS

Signed-off-by: Daniel Dragomir <Daniel.Dragomir@windriver.com>