kernel-devsrc: Add missing files for "modules_prepare" target
Starting with commit 9af0f1a46bbb6ad9ee8b35957251f4aa826b023f
("kernel-devsrc: restructure for out of tree (and on target) module
builds") from poky repo, the way how kernel-devsrc package is
created was changed.

The old kernel-devsrc package starts with a full copy of the kernel
source and then starts to strip out elements that are not required.

This restructuring of the package uses an approach similar to other
distros, where the kernel-devsrc package is for building against the
running kernel and uses a curated set of copied infrastructure, versus
a mass copy of the entire kernel.

This patch add in kernel-devsrc some missing headers and other files
required to perform the "modules_prepare" target.

Signed-off-by: Daniel Dragomir <Daniel.Dragomir@windriver.com>