linux: Refactoring kernel recipes to avoid code duplication
The main reason for this refactoring of the kernel recipes is to
avoid setting SRCREV and LINUX_VERSION with same value for both
debug and default version of the kernels (standard and preempt-rt).

Now, the only place where we'll make changes for a kernel update
will be linux-yocto.inc and linux.yocto.inc where we can set
LINUX_VERSION for standard and preempt-rt only once and SRCREV for
standard and preempt-rt only once. And also the patches will be
updated from the same inc file.

This change don't bring functional changes.

Signed-off-by: Daniel Dragomir <Daniel.Dragomir@windriver.com>