busybox-udhcpc: Handle case when rootfs is mounted as overlayfs
The custom axxia rc.local script run on the end of the boot
process the udhcpc command for the main network interface eth0
in order to populate resolv.conf with nameservers from dhcp lease.
The dpcp client calls the /etc/udhcpc.d/50default script which
check first (in root_is_nfs function) if the interface is used for
nfsroot and skip ip commands which can break nfs conectivity.

This patch is handling the case when the rootfs is mounted as
overlayfs by overwriting the root_is_nfs function: extracts the
lowerdir directory of the overlay (the read-only rootfs) and check
if this is mounted via nfs|smbfs|ncp|coda and takes actions to not
break conectivity.

Otherwise the board will freeze on the end of the boot and after
timeout will reboot.

Signed-off-by: Daniel Dragomir <Daniel.Dragomir@windriver.com>