bootlogd: add support for custom ttyAMA console
Bootlogd daemon (part of sysvinit package) runs in the background
and copies all strings sent to the /dev/console device to a logfile.
It opens /dev/console and finds out what the real console is with an
ioctl() if TIOCCONS is available. Otherwise bootlogd tries to parse
/proc/cmdline for console=  kernel command line arguments.

The the bootlogd documentation there is a known bug (or uncoveredcase):
If that syntax is ever changed by the kernel, or a console type is
used that bootlogd does not know about then bootlogd will not work.

Axxia Gen2 board are using an ARM custom console named ttyAMA which is
not known by bootlogd. That's the reason why the messages from boot are
just redirected to the log file (/var/log/boot) but not shown up in
the console.

This patch adds support for the custom ttyAMA console in bootlogd code.
The boot messages are now shown on console and also in log file.

B0230 - Corner Case -- Console Output Missing from rc.local

Signed-off-by: Daniel Dragomir <Daniel.Dragomir@windriver.com>